+++ ui/main/game/galactic_war/cards/gwaio_upgrade_radarjammer.js
@@ -1,0 +1,47 @@
+define([
+  "coui://ui/mods/com.pa.quitch.gwaioverhaul/shared/cards.js",
+  "coui://ui/mods/com.pa.quitch.gwaioverhaul/shared/units.js",
+], function (gwoCard, gwoUnit) {
+  return {
+    visible: _.constant(true),
+    describe: _.constant(
+      "!LOC:Radar Jamming Station Upgrade Tech doubles the jamming and radar radius of the radar jammer."
+    ),
+    summarize: _.constant("!LOC:Radar Jamming Station Upgrade Tech"),
+    icon: _.constant(
+      "coui://ui/mods/com.pa.quitch.gwaioverhaul/gw_play/img/tech/gwc_intelligence_fabrication_upgrade.png"
+    ),
+    audio: function () {
+      return {
+        found: "/VO/Computer/gw/board_tech_available_efficiency",
+      };
+    },
+    getContext: gwoCard.getContext,
+    deal: function (system, context, inventory) {
+      var chance = 0;
+      if (gwoCard.hasUnit(inventory.units(), gwoUnit.radarJammingStation)) {
+        chance = 60;
+      }
+      return { chance: chance };
+    },
+    buff: function (inventory) {
+      inventory.addMods([
+        {
+          file: gwoUnit.radarAdvanced,
+          path: "recon.observer.items.1.radius",
+          op: "multiply",
+          value: 2,
+        },
+        {
+          file: gwoUnit.radarAdvanced,
+          path: "recon.observer.items.3.radius",
+          op: "multiply",
+          value: 2,
+        },
+      ]);
+    },
+    dull: function () {
+      // empty
+    },
+  };
+});